# swagger.ori.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Service Profil",
        "version": "1.0.0",
        "title": "Service Profil",
        "contact": {
            "name": "Racka Pratama",
            "url": "@rackapratamaps",
            "email": "racka@indonesian-aerospace.com"
        }
    },
    "host": "10.1.94.235:8002",
    "basePath": "/",
...
    "paths": {
        "/aplikasi": {
            "get": {
                "tags": [
                    "Aplikasi"
                ],
                "summary": "Fungsi untuk menampilkan aplikasi",
                "operationId": "getAllUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "aktif",
                        "in": "query",
                        "description": "Filter status aktif, kosongkan untuk mengambil semua aplikasi, aktif = 1 jika hanya yang aktif, aktif = 0 jika semua yang nonaktif",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false
                    },
                    {
                        "name": "semua",
                        "in": "query",
                        "description": "Filter semua aplikasi, semua = 1 atau kosongkan untuk mengambil list aplikasi termasuk \"Semua Aplikasi\" dan SAP, semua = 0 untuk mengambil list aplikasi tanpa \"Semua Aplikasi\" dan SAP",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Response Aplikasi"
                            }
                        }
                    },
                    "404": {
                        "description": "Aplikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
        "/aplikasi/tambah": {
            "post": {
                "tags": [
                    "Aplikasi"
                ],
                "summary": "Fungsi untuk menambah aplikasi",
                "operationId": "saveUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Aplikasi",
                        "description": "Format JSON untuk request aplikasi",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Aplikasi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Aplikasi"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "409": {
                        "description": "Aplikasi dengan kode {kode} sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
        "/menu": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Fungsi untuk menampilkan menu",
                "operationId": "getAllUsingGET_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "aktif",
                        "in": "query",
                        "description": "Filter status aktif, kosongkan untuk mengambil semua menu, aktif = 1 jika hanya yang aktif, aktif = 0 jika semua yang nonaktif",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false
                    },
                    {
                        "name": "sidebar",
                        "in": "query",
                        "description": "Filter kode sidebar, kosongkan untuk mengambil semua menu, sidebar = 1 jika hanya yang menu di sidebar, sidebar = 0 jika semua yang bukan di sidebar",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Response Menu"
                            }
                        }
                    },
                    "404": {
                        "description": "Menu tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
        "/menu/hapus": {
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "Fungsi untuk menghapus menu",
                "operationId": "deleteUsingDELETE_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Id",
                        "description": "Format JSON untuk request id",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Umum"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "404": {
                        "description": "Tidak ditemukan Menu dengan Id {id}"
                    },
                    "500": {
                        "description": "Hapus semua entitas anak dari Menu dengan Id {id}"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
        "/menu/tambah": {
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Fungsi untuk menambah menu",
                "operationId": "saveUsingPOST_1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Menu",
                        "description": "Format JSON untuk request menu",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Menu"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "404": {
                        "description": "Tidak ditemukan Menu Induk dengan Id {id}"
                    },
                    "409": {
                        "description": "Menu dengan kode {kode} sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
        "/menu/ubah": {
            "put": {
                "tags": [
                    "Menu"
                ],
                "summary": "Fungsi untuk mengubah menu",
                "operationId": "updateUsingPUT_1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Menu",
                        "description": "Format JSON untuk request menu",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Menu"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "404": {
                        "description": "Tidak ditemukan Menu dengan Id {id}"
                    },
                    "409": {
                        "description": "Menu dengan kode {kode} sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
        "/menu/{id}": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Fungsi untuk menampilkan menu",
                "description": "Berdasarkan id menu",
                "operationId": "getOneUsingGET_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id menu",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Menu"
                        }
                    },
                    "400": {
                        "description": "Id tidak valid"
                    },
                    "404": {
                        "description": "Menu tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}